+++ src/containers/Main/game.tsx
@@ -1,56 +1,70 @@
 import { MAX_POS } from '@assets/constant';
+import { Box, Button, Link, Text, useToast } from '@chakra-ui/react';
 import useMoveTile from '@hooks/useMoveTile';
-import { getInitialTileList } from '@utils/tile';
-import { times } from 'lodash';
-import { useState } from 'react';
+import { initialBoardSetting } from '@utils/tile2';
+import { useEffect, useState } from 'react';
 import { GameContainer, GridCell, GridContainer, InnerText } from './style';
 
-type Item = {
-  xPos: number;
-  yPos: number;
-  value: number;
-  isNew: boolean;
-  isMerged: boolean;
-  isDisabled: boolean;
-};
-
 const Game = () => {
-  const [tileList, setTileList] = useState<Item[]>(getInitialTileList);
-  useMoveTile({ tileList, setTileList });
+  const [board, setBoard] = useState<number[][]>(
+    initialBoardSetting(
+      Array.from(new Array(MAX_POS), () => new Array(MAX_POS).fill(0)),
+    ),
+  );
+  const [isGameOver, setIsGameOver] = useState(false);
+  const toast = useToast();
+  const id = 'test';
+  useMoveTile({ board, setBoard, setIsGameOver });
+
+  useEffect(() => {
+    if (isGameOver) {
+      if (!toast.isActive(id)) {
+        toast({
+          id,
+          status: 'error',
+          duration: null,
+          position: 'top',
+          render: () => (
+            <>
+              <Box bg="tomato" w="100%" p={4} color="white">
+                <Text fontSize="lg" fontWeight="bold" mb={2}>
+                  Game Over
+                </Text>
+                <Text mb={4}>더 이상 이동할 수 있는 타일이 없습니다.</Text>
+                <Link href="/">
+                  <Button colorScheme="red">New Game</Button>
+                </Link>
+              </Box>
+            </>
+          ),
+        });
+      }
+    }
+    return () => {
+      setIsGameOver(false);
+    };
+  }, [isGameOver, toast]);
+
   return (
     <>
       <GameContainer>
         <GridContainer>
-          {/* 카드 기본 틀 만들기 */}
-          {times(MAX_POS, (y) =>
-            times(MAX_POS, (x) => (
-              <GridCell
-                key={`cell-${y + 1}-${x + 1}`}
-                $xPos={y + 1}
-                $yPos={x + 1}
-              >
-                <InnerText>
-                  ({y + 1},{x + 1})
-                </InnerText>
-              </GridCell>
-            )),
+          {/* board map으로 돌기 */}
+          {board.map((row, rowIndex) =>
+            row.map((value, colIndex) => {
+              return (
+                <GridCell
+                  key={`tile-${rowIndex}-${colIndex}-${value}`}
+                  $xPos={rowIndex + 1}
+                  $yPos={colIndex + 1}
+                >
+                  <InnerText $color={value}>
+                    {value !== 0 ? value : undefined}
+                  </InnerText>
+                </GridCell>
+              );
+            }),
           )}
-
-          {/* tile List map으로 돌기 */}
-          {tileList.map((item) => {
-            return (
-              <GridCell
-                key={`tile-${item.yPos}-${item.xPos}-${item.value}-${item.isMerged}`}
-                $xPos={item.yPos}
-                $yPos={item.xPos}
-              >
-                <InnerText $color={item.value}>
-                  ({item.yPos},{item.xPos})<br />
-                  {item.value}
-                </InnerText>
-              </GridCell>
-            );
-          })}
         </GridContainer>
       </GameContainer>
     </>
